feat: moved over solution to spring boot starter
