+++ clients/camunda-spring-boot-starter/src/main/java/io/camunda/client/jobhandling/result/DefaultResultProcessorStrategy.java
@@ -16,8 +16,8 @@
 package io.camunda.client.jobhandling.result;
 
 import io.camunda.client.CamundaClient;
-import io.camunda.client.api.command.CompleteJobCommandStep1.ResultFunction;
 import io.camunda.client.bean.MethodInfo;
+import io.camunda.client.jobhandling.UserTaskResultFunction;
 
 public class DefaultResultProcessorStrategy implements ResultProcessorStrategy {
   private final CamundaClient camundaClient;
@@ -36,7 +36,7 @@ public class DefaultResultProcessorStrategy implements ResultProcessorStrategy {
   @Override
   public ResultProcessor createProcessor(final MethodInfo methodInfo) {
     if (methodInfo.getReturnType() != null
-        && methodInfo.getReturnType().equals(ResultFunction.class)) {
+        && methodInfo.getReturnType().equals(UserTaskResultFunction.class)) {
       return new ResultFunctionResultProcessor();
     }
     return new DefaultResultProcessor(
